Refactored XLinkLog
